==> examples/0-Get_campaign_info.ipynb <==
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
import time

import sys, os
sys.path.insert(0, os.path.abspath('../'))
#sys.path.insert(0, '/Users/jorge/git/crowdfunding-prophet/')

from cwdprophet.campaign import *
import pandas as pd

Let's quickly try a webdriver by loading www.google.com

chrome_options = Options()
chrome_options.headless = False
#chrome_options.add_argument("--headless")

driver = webdriver.Chrome(options=chrome_options)
driver.get('https://www.google.com')
time.sleep(2)
driver.quit()

### Campaign Info

#url = 'https://www.crowdfunder.co.uk/the-community-waffle-house'
url = 'https://www.crowdfunder.co.uk/save-the-new-inn-norton-lindsey'
campaign = load_campaign(campaign_url=url,browser=None,quit_browser=False)
pledges, dates, names = campaign.pledges()
campaign.browser.quit()

campaign.report()

df = pd.DataFrame({'date': dates,
                   'supporter name': names,
                   'pledge': pledges})
df
