==> examples/2-Campaign_Info.ipynb <==
from selenium import webdriver
from bs4 import BeautifulSoup
import time
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd

import sys, os
sys.path.insert(0, os.path.abspath('../'))
from cwdprophet.campaign import *

df = pd.read_csv('Campaigns.csv',index_col=0)
df.head()

# Let's initialize lists to save all the campaign details including:
# creator, location, deadline, duration, 
campaigns_creators = [];
campaigns_locs = [];
campaigns_deadlines = [];
campaigns_lengths = [];

# FB page, Twitter page
campaigns_fb = [];
campaigns_tw = [];

# And the number of comments, updates and donnations made
campaigns_Nupdates = [];
campaigns_Ncomments = [];
campaigns_Ninvestors = [];


browser = webdriver.Chrome()

for url in df['URL'].tolist()[750:1000]:
    
    campaign = load_campaign(campaign_url=url,browser=browser,quit_browser=False)
    
    campaigns_creators.append(campaign.creator())
    campaigns_locs.append(campaign.location())
    campaigns_deadlines.append(campaign.deadline())
    campaigns_lengths.append(campaign.duration())
    campaigns_fb.append(campaign.fbpage())
    campaigns_tw.append(campaign.twpage())
    campaigns_Nupdates.append(campaign.Nupdates())
    campaigns_Ncomments.append(campaign.Ncomments())
    campaigns_Ninvestors.append(campaign.Ninvestors())
        
    if url in df['URL'].tolist()[0:3]:
        campaign.report()
        print()

    time.sleep(1)

campaign.browser.quit()

new_campaigns_locs = [x if x is not '' else None for x in campaigns_locs]

df['Author'] = campaigns_creators
df['Location'] = new_campaigns_locs
df['Deadline'] = campaigns_deadlines
df['Duration'] = campaigns_lengths
df['Nupdates'] = campaigns_Nupdates
df['Ncomments'] = campaigns_Ncomments
df['Nsupporters'] = campaigns_Ninvestors
df['Facebook'] = campaigns_fb
df['Twitter'] = campaigns_tw

len(df)

df.to_csv('Community.csv')
pd.options.display.float_format = '{:.0f}'.format
df.head()
