==> examples/3-History_of_donations_per_campaign.ipynb <==
from selenium import webdriver
from bs4 import BeautifulSoup
import time
from matplotlib import pyplot as plt
import numpy as np
import pandas as pd

import sys, os
sys.path.insert(0, os.path.abspath('../'))
from cwdprophet.campaign import *

df = pd.read_csv('Community.csv',index_col=0)
df.head()

# Let's initialize lists to save all the pledges, with their corresponding
# dates and supporter names.
total_pledges = []
total_dates = []
total_names = []

browser = webdriver.Chrome()
for url in df['URL'].tolist()[0:3]:
    
    campaign = load_campaign(campaign_url=url,browser=browser,close_browser=False,quit_browser=False)
    pledges, dates, names = campaign.pledges()

    total_pledges.append(pledges)
    total_dates.append(dates)
    total_names.append(names)
    
campaign.browser.quit()

# Check how many campaigns were scraped
len(total_pledges), len(total_dates), len(total_names)

total_pledges[2]
